+++ lib/browser/chrome.rb
@@ -31,6 +31,7 @@ module Browser
         !huawei_browser? &&
         !miui_browser? &&
         !maxthon? &&
+        !mail_master? &&
         !google_search_app?
     end
   end